init: port from Neovim Haml files
My goal is to create a Haml template grammar for Treesitter in order
to get syntax highlighting in the Zed editor.

I am coming from Neovim, which has nice Haml support.

I gave these files to ChatGPT 4o and asked it to write the Treesitter
grammar based on their contents:

```
/opt/homebrew/Cellar/neovim/0.10.1/share/nvim/runtime/indent/haml.vim
/opt/homebrew/Cellar/neovim/0.10.1/share/nvim/runtime/syntax/haml.vim
/opt/homebrew/Cellar/neovim/0.10.1/share/nvim/runtime/syntax/html.vim
/opt/homebrew/Cellar/neovim/0.10.1/share/nvim/runtime/syntax/javascript.vim
/opt/homebrew/Cellar/neovim/0.10.1/share/nvim/runtime/syntax/ruby.vim
```